Update files, Add change, edit, deleteAll, delete
